fix incorrect substitution of fuzzball URI object with java.net.URI
the groovy code generation code in the openapi codegen tool
subclasses the abstract java generation class which in turn uses a hardcoded
mapping of URI -> java.net.URI. However, the fuzzball OpenAPI spec defines
its own URI object. This commit is a stopgap solution to this problem. In the
longterm we need a more robust mechanism for doing this but for now this
will do.